feat : deleteImage 로직 추가

+++ spring-backend/src/main/java/com/linkey/core/domain/dto/request/ReqUpdateImageDto.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class ReqUpdateImageDto {
-    private int imgId;
+    private long imgId;
     private SingleProject project;
     private SingleSprint sprint;
